+++ src/main/java/software/works/fizzbuzz/rule/FunctionTypes.java
@@ -1,14 +1,15 @@
 package software.works.fizzbuzz.rule;
 
+import java.math.BigInteger;
 import java.util.function.BiFunction;
 import java.util.function.BiPredicate;
 import java.util.function.Predicate;
 
 public interface FunctionTypes {
 
-    static interface PropertyPredicate extends BiPredicate<Long, Integer> {}
+    static interface PropertyPredicate extends BiPredicate<BigInteger, BigInteger> {}
 
-    static interface WordOccurrencesFunction extends BiFunction<Long, WordPropertyPair, String> {}
+    static interface WordOccurrencesFunction extends BiFunction<BigInteger, WordPropertyPair, String> {}
 
-    static interface ValuePredicate extends Predicate<Long> {}
+    static interface ValuePredicate extends Predicate<BigInteger> {}
 }
